Move sari_score.py to parent folder (and mimic same for samsa)
