+++ easse/samsa/__init__.py
@@ -1,0 +1,1 @@
+from easse.samsa.samsa_score import samsa_corpus
